[BE] refactor : Enum 이름 변경, final 키워드 추가, long primitive 타입으로 변경
