+++ backend/src/main/java/booki_today/domain/assemble/AssembleType.java
@@ -1,4 +1,4 @@
-package booki_today.domain.created_booki;
+package booki_today.domain.assemble;
 
 public enum AssembleType {
