+++ backend/src/main/java/booki_today/domain/category/CategoryType.java
@@ -2,5 +2,5 @@ package booki_today.domain.category;
 
 public enum CategoryType {
 
-    UNIVERSITY, CLUB, DEPARTMENT
+    UNIVERSITY, DEPARTMENT, CLUB
 }
