+++ backend/src/main/java/booki_today/domain/participated_assemble/ParticipatedStatus.java
@@ -1,4 +1,4 @@
-package booki_today.domain.participated_booki;
+package booki_today.domain.participated_assemble;
 
 public enum ParticipatedStatus {
